+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,15 @@
 # TBD
 
+# 0.17.0
+### Features
+* Added `PauseService` and `UnpauseService` to `KurtosisBackend`
+* Added docker implementation of `PauseService` and `UnpauseService`
+* Added Kubernetes implementation of engine functions in kubernetes backend
+
+### Breaking Changes
+* Added an extra `enclaveDataVolumeDirpath` to `KurtosisBackend.CreateAPIContainer`
+    * Users should pass in the location where the enclave data volume should be mounted
+
 # 0.16.0
 ### Removals
 * Removed `files_artifact.FilesArtifactID` because it was a duplicate of `serivce.FilesArtifactID`